Preparing login.php
The function was going to come, but i noticed that the form was split in two. This meant 1. a new grid-template needed to be made for the form 2. the animations needed to be fixed 3. the elements needed to be fitted to the new form. Now it's ready to handle actual PHP functions (i can't promise the next commit apparently).