Quick team changing
Added the ability to quickly change teams, this doesn't refresh or require a manual page refresh, nor does it need the user to restart their session (log back in). The code updates the database alongside the UI. As of this commit, the debugging information is still on the panel, with the controls.