URL encoding for white-space in  variable
Fixes #3 - this commit allows the  variable to have valid white-space. This was successfully done with a fairly simple PHP function 'urlencode' to encode special characters into the URL so that it can handle the white-spaces. As of this commit, PHP automatically decodes the URL when using , but incase of probable change the 'urldecode' function was included anyway. At the moment 'event.php' is essentially just a testing page, showing the variable carried by the URL.